+++ dsl/src/commonMain/kotlin/it/unibo/collektive/aggregate/ops/AggregateOps.kt
@@ -1,5 +1,6 @@
-package it.unibo.collektive.aggregate
+package it.unibo.collektive.aggregate.ops
 
+import it.unibo.collektive.aggregate.AggregateContext
 import it.unibo.collektive.field.Field
 
 /**
