+++ plugin/compiler-plugin/src/main/kotlin/it/unibo/collektive/utils/common/AggregateFunctionNames.kt
@@ -3,6 +3,6 @@ package it.unibo.collektive.utils.common
 object AggregateFunctionNames {
     const val ALIGNED_ON_FUNCTION = "alignedOn"
     const val AGGREGATE_FUNCTION = "aggregate"
-    const val ALIGNED_ON_FUNCTION_FQ_NAME = "it.unibo.collektive.AggregateContext.alignedOn"
-    const val AGGREGATE_CONTEXT_CLASS = "it.unibo.collektive.AggregateContext"
+    const val ALIGNED_ON_FUNCTION_FQ_NAME = "it.unibo.collektive.aggregate.AggregateContext.alignedOn"
+    const val AGGREGATE_CONTEXT_CLASS = "it.unibo.collektive.aggregate.AggregateContext"
 }
